export classes to modules to tidy up code
