+++ components/shooter.py
@@ -1,0 +1,13 @@
+import pygame
+pygame.init()
+
+class Shooter(pygame.sprite.Sprite):
+	def __init__(self, x, y):
+		pygame.sprite.Sprite.__init__(self)
+		self.image = pygame.image.load('img/R1E.png')
+		self.rect = self.image.get_rect()
+		self.rect.x = x
+		self.rect.y = y
+
+	def update(self):
+		pass
